scala 2.13.7, sbt 1.5.5, update dependencies, adapt code, Xsource:3

+++ src/main/scala/jackdaw/key/KeyDetector.scala
@@ -32,7 +32,7 @@ object KeyDetector {
 		*/
 		val audio	= new Audio {
 			def frameCount	= sample.frameCount
-			def frameRate	= sample.frameRate
+			def frameRate	= sample.frameRate.toDouble
 			def get(frame:Int):Float	= mono get frame
 		}
